# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Project API",
        "description": "Badminton Project API Documentation",
        "contact": {
            "email": "admin@badminton.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local Server"
        },
        {
            "url": "https://api.badminton.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/getAllVenue": {
            "get": {
                "tags": [
                    "Venue"
                ],
                "summary": "Lấy danh sách tất cả sân",
                "description": "API lấy danh sách tất cả sân badminton",
                "operationId": "getVenueList",
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/getMyVenues": {
            "get": {
                "tags": [
                    "Venue"
                ],
                "summary": "Lấy danh sách sân của chủ sân",
                "description": "API lấy danh sách sân badminton của chủ sân",
                "operationId": "getMyVenues",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID của chủ sân",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/getVenueDetail/{id}": {
            "get": {
                "tags": [
                    "Venue"
                ],
                "summary": "Lấy thông tin chi tiết sân",
                "description": "API lấy thông tin chi tiết sân badminton theo ID",
                "operationId": "getVenueDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của sân",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thông tin thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy sân"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}